After some mail exchange with Roberto I decided to move out the CppPreprocessor and CppDocument code out of cplusplus/preprocessor. This is the first step for that. The point is that the implementation of those classes depend on the actual goal (Incremental updates in case of Qt Creator and fact extraction of existing documents in our case).
The Document class in cppmodel is a read only class which build by (its friend
class) CppPreprocessor. Also, the CppPreprocessor class now can distinguis
between local or quote includes and global includes. It will add a warning to
the document when a local include only could be found in a global include
directory.


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/quality/krazy2@938812 283d02a7-25f6-0310-bc7c-ecb5cbfe19da