Attempt to fix lxr links for koffice in bundled-apps.
This should work for koffice. Please check the links in a day or two
However, removing bundled-apps from the the  path doesn't work for
the other module in bundled-apps (i.e. kdepim-enterprise). I'm not
sure if source for that are on lxr anyway.

CCMAIL: staniek@kde.org
CCMAIL: winter@kde.org


git-svn-id: svn+ssh://svn.kde.org/home/kde/trunk/quality/krazy2@1089320 283d02a7-25f6-0310-bc7c-ecb5cbfe19da