Fix not processing the line after a possible problem.
The line number was increased after finding a possible problem, but that
shouldn't be done, as the line number will be increased again when
getting the next line to be processed. Thus, the line number was
increased twice after finding a possible problem, and the line after the
one with the possible problem was not processed.

Signed-off-by: Allen Winter <allen.winter@kdab.com>