Search for "private:" only in the declaration of the class.
Before, a fixed amount of lines to search back for "private:" was used.
Now, it is searched back from the current line to the start of the
declaration of the class. This prevents finding the "private:" of
another class also declared in the same header, or not finding the
"private:" in headers with a lot of private methods.

Signed-off-by: Allen Winter <allen.winter@kdab.com>