Fix not taking into account different visibilities.
Before, "private:" was searched from the current method up to the
declaration of the class. Thus, if a group of private methods was
declared, and then a group of public methods was declared, the public
methods were treated as private. Now, it is searched for the previous
"public:", "protected:" and "private:" declarations, and a method is
treated as private only if "private:" is the nearest previous visibility
declaration.

Signed-off-by: Allen Winter <allen.winter@kdab.com>