Fix limit to search back.
The search is performed on n lines before the given l line. Thus, n can
be, at the most, l. Otherwise, if n is larger than l, at some point
$l-$i will become a negative number, due to i being lesser or equal than
n. This causes that after the search is performed on the line 0, the
next line that will be searched will be the last line, then the line
before the last line and so on. Thus, a "private:" in the lower part of
a header could be erroneously found if the public method being checked
was in the upper part of the file.

Signed-off-by: Allen Winter <allen.winter@kdab.com>